requirements.txt
arrow==0.12.1
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
binaryornot==0.4.4
certifi==2018.8.13
chardet==3.0.4
click==6.7
colorama==0.3.9
cookiecutter==1.6.0
coverage==4.5.1
coveralls==1.3.0
docopt==0.6.2
Flask==1.0.2
future==0.16.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jinja2-time==0.2.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.7.1
poyo==0.4.1
py==1.5.4
pylint==2.1.1
pytest==3.7.2
pytest-cov==2.5.1
python-dateutil==2.7.3
PyJWT==1.6.4
PyYAML==3.13
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
Werkzeug==0.14.1
whichcraft==0.4.1
wrapt==1.10.11
